Refactored filter, forms, tables and views and the test framework
