+++ netbox_dns/filters/record.py
@@ -2,42 +2,8 @@ import django_filters
 from django.db.models import Q
 
 from netbox.filtersets import NetBoxModelFilterSet
-from .models import NameServer, Record, Zone
 
-
-class ZoneFilter(NetBoxModelFilterSet):
-    """Filter capabilities for Zone instances."""
-
-    q = django_filters.CharFilter(
-        method="search",
-        label="Search",
-    )
-    status = django_filters.ChoiceFilter(
-        choices=Zone.CHOICES,
-    )
-
-    class Meta:
-        model = Zone
-        fields = ("name", "status", "nameservers")
-
-    def search(self, queryset, name, value):
-        """Perform the filtered search."""
-        if not value.strip():
-            return queryset
-        qs_filter = Q(name__icontains=value) | Q(status__icontains=value)
-        return queryset.filter(qs_filter)
-
-
-class NameServerFilter(NetBoxModelFilterSet):
-    """Filter capabilities for NameServer instances."""
-
-    name = django_filters.CharFilter(
-        lookup_expr="icontains",
-    )
-
-    class Meta:
-        model = NameServer
-        fields = ("name",)
+from netbox_dns.models import Record, Zone
 
     def search(self, queryset, name, value):
         """Perform the filtered search."""
