+++ netbox_dns/tables/__init__.py
@@ -1,0 +1,3 @@
+from .nameserver import *
+from .record import *
+from .zone import *
