+++ netbox_dns/views/__init__.py
@@ -1,0 +1,3 @@
+from .nameserver import *
+from .record import *
+from .zone import *
